Move most of tests/slg to use the generic test infra
