+++ tests/test/mod.rs
@@ -102,7 +102,7 @@ macro_rules! test {
               @unparsed_goals[$($unparsed_goals)*])
     };
 
-    // goal { G } yields[C1] { "Y1" } yields[C2] { "Y2" } -- test that solver C1 yields Y1
+    // goal { G } yields_all[C1] { "Y1" } yields_all[C2] { "Y2" } -- test that solver C1 yields Y1
     // and C2 yields Y2
     //
     // Annoyingly, to avoid getting a parsing ambiguity error, we have
@@ -110,22 +110,57 @@ macro_rules! test {
     // (this rule) for the last goal in the list (next rule). There
     // might be a more elegant fix than copy-and-paste but this works.
     (@program[$program:tt] @parsed_goals[$($parsed_goals:tt)*] @unparsed_goals[
-        goal $goal:tt $(yields[$($C:expr),+] { $expected:expr })*
+        goal $goal:tt yields_all[$C:expr] { $($expected:expr),* }
             goal $($unparsed_goals:tt)*
     ]) => {
         test!(@program[$program]
-              @parsed_goals[$($parsed_goals)*
-                            $($((stringify!($goal), $C, TestGoal::All(vec![$expected])))+)+]
+              @parsed_goals[
+                  $($parsed_goals)*
+                      (stringify!($goal), $C, TestGoal::All(vec![$($expected),*]))
+              ]
               @unparsed_goals[goal $($unparsed_goals)*])
     };
 
     // same as above, but for the final goal in the list.
     (@program[$program:tt] @parsed_goals[$($parsed_goals:tt)*] @unparsed_goals[
-        goal $goal:tt $(yields[$($C:expr),+] { $expected:expr })*
+        goal $goal:tt yields_all[$C:expr] { $($expected:expr),* }
     ]) => {
         test!(@program[$program]
-              @parsed_goals[$($parsed_goals)*
-                            $($((stringify!($goal), $C, TestGoal::All(vec![$expected])))+)+]
+              @parsed_goals[
+                  $($parsed_goals)*
+                      (stringify!($goal), $C, TestGoal::All(vec![$($expected),*]))
+                ]
+              @unparsed_goals[])
+    };
+
+    // goal { G } yields_first[C1] { "Y1" } yields_first[C2] { "Y2" } -- test that solver C1 yields Y1
+    // and C2 yields Y2
+    //
+    // Annoyingly, to avoid getting a parsing ambiguity error, we have
+    // to distinguish the case where there are other goals to come
+    // (this rule) for the last goal in the list (next rule). There
+    // might be a more elegant fix than copy-and-paste but this works.
+    (@program[$program:tt] @parsed_goals[$($parsed_goals:tt)*] @unparsed_goals[
+        goal $goal:tt yields_first[$C:expr] { $($expected:expr),* }
+            goal $($unparsed_goals:tt)*
+    ]) => {
+        test!(@program[$program]
+              @parsed_goals[
+                  $($parsed_goals)*
+                      (stringify!($goal), $C, TestGoal::First(vec![$($expected),*]))
+              ]
+              @unparsed_goals[goal $($unparsed_goals)*])
+    };
+
+    // same as above, but for the final goal in the list.
+    (@program[$program:tt] @parsed_goals[$($parsed_goals:tt)*] @unparsed_goals[
+        goal $goal:tt yields_first[$C:expr] { $($expected:expr),* }
+    ]) => {
+        test!(@program[$program]
+              @parsed_goals[
+                  $($parsed_goals)*
+                      (stringify!($goal), $C, TestGoal::First(vec![$($expected),*]))
+                ]
               @unparsed_goals[])
     };
 }
@@ -211,3 +246,4 @@ mod negation;
 mod projection;
 mod unify;
 mod wf_goals;
+mod slgext;
