+++ tests/test/slg.rs
@@ -84,153 +84,6 @@ fn solve_goal_fixed_num_answers(program_text: &str, goals: Vec<(usize, usize, &s
 }
 
 #[test]
-fn basic() {
-    test! {
-        program {
-            trait Sized { }
-
-            struct i32 { }
-            impl Sized for i32 { }
-        }
-
-        goal {
-            forall<T> { if (T: Sized) { T: Sized } }
-        } first 2 with max 10 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [],
-                            constraints: []
-                        },
-                        binders: []
-                    },
-                    ambiguous: false
-                }
-            ]"
-        }
-    }
-}
-
-#[test]
-fn breadth_first() {
-    test! {
-        program {
-            trait Sized { }
-
-            struct i32 { }
-            impl Sized for i32 { }
-
-            struct Vec<T> { }
-            impl<T> Sized for Vec<T> where T: Sized { }
-
-            struct Slice<T> { }
-            impl<T> Sized for Slice<T> where T: Sized { }
-        }
-
-        goal {
-            exists<T> { T: Sized }
-        } first 5 with max 10 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [?0 := i32],
-                            constraints: []
-                        },
-                        binders: []
-                    },
-                    ambiguous: false
-                },
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [?0 := Vec<i32>],
-                            constraints: []
-                        },
-                        binders: []
-                    },
-                    ambiguous: false
-                },
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [?0 := Slice<i32>],
-                            constraints: []
-                        },
-                        binders: []
-                    },
-                    ambiguous: false
-                },
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [?0 := Vec<Vec<i32>>],
-                            constraints: []
-                        },
-                        binders: []
-                    },
-                    ambiguous: false
-                },
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [?0 := Slice<Vec<i32>>],
-                            constraints: []
-                        },
-                        binders: []
-                    },
-                    ambiguous: false
-                }
-            ]"
-        }
-    }
-}
-
-#[test]
-fn infinite_recursion() {
-    test! {
-        program {
-            trait A { }
-            trait B { }
-            trait C { }
-            trait D { }
-
-            struct Vec<T> { }
-            impl<T> A for Vec<T> where T: B { }
-            impl<T> B for Vec<T> where T: C { }
-            impl<T> C for Vec<T> where T: D { }
-            impl<T> D for Vec<T> where T: A { }
-        }
-
-        goal {
-            exists<T> { T: A }
-        } first 5 with max 10 {
-            r"[]"
-        }
-    }
-}
-
-/// Make sure we don't get a stack overflow or other badness for this
-/// test from scalexm.
-#[test]
-fn subgoal_abstraction() {
-    test! {
-        program {
-            trait Foo { }
-            struct Box<T> { }
-            impl<T> Foo for T where Box<T>: Foo { }
-        }
-
-        goal {
-            exists<T> { T: Foo }
-        } first 5 with max 50 {
-            r"[]"
-        }
-    }
-}
-
-#[test]
 fn flounder() {
     test! {
         program {
@@ -250,6 +103,7 @@ fn flounder() {
 
 // Test that, when solving `?T: Sized`, we only wind up pulling a few
 // answers before we stop.
+// FIXME: This is basically the same as `breadth_first`. Is it testing something different?
 #[test]
 fn only_draw_so_many() {
     test! {
@@ -270,55 +124,6 @@ fn only_draw_so_many() {
             exists<T> { T: Sized }
         } fixed 2 with max 10 {
             "Some(Ambig(Unknown))"
-        }
-    }
-}
-
-#[test]
-fn only_draw_so_many_blow_up() {
-    test! {
-        program {
-            trait Sized { }
-            trait Foo { }
-
-            struct Vec<T> { }
-            impl<T> Sized for Vec<T> where T: Sized { }
-            impl<T> Foo for Vec<T> where T: Sized { }
-
-            struct i32 { }
-            impl Sized for i32 { }
-
-            struct Slice<T> { }
-            impl<T> Sized for Slice<T> where T: Sized { }
-        }
-
-        goal {
-            exists<T> { T: Foo }
-        } fixed 2 with max 10 {
-            "Some(Ambig(Definite(Canonical { value: [?0 := Vec<^0>], binders: [Ty(U0)] })))"
-        }
-    }
-}
-
-/// Here, P and Q depend on one another through a negative loop.
-#[test]
-#[should_panic(expected = "negative cycle")]
-fn negative_loop() {
-    test! {
-        program {
-            trait P { }
-            trait Q { }
-            struct u32 { }
-
-            forall<> { u32: P if not { u32: Q } }
-            forall<> { u32: Q if not { u32: P } }
-        }
-
-        goal {
-            u32: P
-        } first 5 with max 3 {
-            // Negative cycle -> panic
-            r""
         }
     }
 }
@@ -420,245 +225,6 @@ fn subgoal_cycle_uninhabited() {
                     ambiguous: false
                 }
             ]"
-        }
-    }
-}
-
-#[test]
-fn subgoal_cycle_inhabited() {
-    test! {
-        program {
-            trait Foo { }
-            struct Box<T> { }
-            struct Vec<T> { }
-            struct u32 { }
-            impl<T> Foo for Box<T> where Box<Vec<T>>: Foo { }
-            impl Foo for u32 { }
-        }
-
-        goal {
-            exists<T> { T: Foo }
-        } first 10 with max 3 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [?0 := u32],
-                            constraints: []
-                        },
-                        binders: []
-                    },
-                    ambiguous: false
-                }
-            ]"
-        }
-    }
-}
-
-#[test]
-fn basic_region_constraint_from_positive_impl() {
-    test! {
-        program {
-            trait Foo { }
-            struct Ref<'a, 'b, T> { }
-            struct u32 { }
-            impl<'x, T> Foo for Ref<'x, 'x, T> { }
-        }
-
-        goal {
-            forall<'a, 'b, T> { Ref<'a, 'b, T>: Foo }
-        } first 10 with max 3 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [],
-                            constraints: [
-                                InEnvironment {
-                                    environment: Env([]),
-                                    goal: '!1_1 == '!1_0
-                                }
-                            ]
-                        },
-                        binders: []
-                    },
-                    ambiguous: false
-                }
-            ]"
-        }
-    }
-}
-
-#[test]
-#[allow(non_snake_case)]
-fn example_2_1_EWFS() {
-    test! {
-        program {
-            trait Edge<B> { }
-            trait TransitiveClosure<B> { }
-            struct a { }
-            struct b { }
-            struct c { }
-
-            forall<> { a: Edge<b> }
-            forall<> { b: Edge<c> }
-            forall<> { b: Edge<a> }
-            forall<X, Y> { X: TransitiveClosure<Y> if X: Edge<Y> }
-            forall<X, Y, Z> { X: TransitiveClosure<Y> if X: Edge<Z>, Z: TransitiveClosure<Y> }
-        }
-
-        goal {
-            exists<V> { a: TransitiveClosure<V> }
-        } first 10 with max 3 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [?0 := b],
-                            constraints: []
-                        },
-                        binders: []
-                    },
-                    ambiguous: false
-                },
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [?0 := c],
-                            constraints: []
-                        },
-                        binders: []
-                    },
-                    ambiguous: false
-                },
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [?0 := a],
-                            constraints: []
-                        },
-                        binders: []
-                    },
-                    ambiguous: false
-                }
-            ]"
-        }
-    }
-}
-
-#[test]
-#[allow(non_snake_case)]
-fn example_2_2_EWFS() {
-    test! {
-        program {
-            trait M { }
-            trait P { }
-            trait Q { }
-
-            struct a { }
-            struct b { }
-            struct c { }
-
-            forall<X> { X: M if not { X: P } }
-            forall<> { a: P }
-            forall<X> { X: P if X: Q }
-            forall<> { b: Q }
-            forall<X> { X: Q if X: P }
-        }
-
-        goal {
-            c: M
-        } first 10 with max 3 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [],
-                            constraints: []
-                        },
-                        binders: []
-                    },
-                    ambiguous: false
-                }
-            ]"
-        }
-    }
-}
-
-#[test]
-#[should_panic(expected = "negative cycle")]
-#[allow(non_snake_case)]
-fn example_2_3_EWFS() {
-    test! {
-        program {
-            trait W { }
-            trait M<A> { }
-            trait P { }
-
-            struct a { }
-            struct b { }
-            struct c { }
-
-            forall<X, Y> { X: W if X: M<Y>, not { Y: W }, Y: P }
-            forall<> { a: M<b> }
-            forall<> { b: M<c> }
-            forall<> { c: M<b> }
-            forall<> { b: P }
-        }
-
-        goal {
-            a: W
-        } first 10 with max 3 {
-            // Negative cycle -> panic
-            r""
-        }
-    }
-}
-
-#[test]
-#[should_panic(expected = "negative cycle")]
-#[allow(non_snake_case)]
-fn example_3_3_EWFS() {
-    test! {
-        program {
-            trait S { }
-            trait P { }
-            trait Q { }
-
-            struct a { }
-
-            forall<> { a: S if not { a: P }, not { a: Q } }
-            forall<> { a: P if not { a: S }, a: Q }
-            forall<> { a: Q if not { a: S }, a: P }
-        }
-
-        goal {
-            a: S
-        } first 10 with max 3 {
-            // Negative cycle -> panic
-            r""
-        }
-    }
-}
-
-/// Here, P is neither true nor false. If it were true, then it would
-/// be false, and so forth.
-#[test]
-#[should_panic(expected = "negative cycle")]
-fn contradiction() {
-    test! {
-        program {
-            trait P { }
-            struct u32 { }
-
-            forall<> { u32: P if not { u32: P } }
-        }
-
-        goal {
-            u32: P
-        } first 10 with max 3 {
-            // Negative cycle -> panic
-            r""
         }
     }
 }
@@ -902,29 +468,6 @@ fn cached_answers_3() {
     }
 }
 
-/// Here, P depends on Q negatively, but Q depends only on itself.
-#[test]
-#[should_panic(expected = "negative cycle")]
-fn negative_answer_ambiguous() {
-    test! {
-        program {
-            trait P { }
-            trait Q { }
-            struct u32 { }
-
-            forall<> { u32: P if not { u32: Q } }
-            forall<> { u32: Q if not { u32: Q } }
-        }
-
-        goal {
-            u32: P
-        } first 10 with max 3 {
-            // Negative cycle -> panic
-            r""
-        }
-    }
-}
-
 #[test]
 fn non_enumerable_traits_direct() {
     test! {
@@ -1046,71 +589,6 @@ fn non_enumerable_traits_double() {
         }
     }
 }
-
-#[test]
-fn non_enumerable_traits_reorder() {
-    test! {
-        program {
-            struct Foo { }
-            struct Bar { }
-
-            #[non_enumerable]
-            trait NonEnumerable { }
-            impl NonEnumerable for Foo { }
-            impl NonEnumerable for Bar { }
-
-            trait Enumerable { }
-            impl Enumerable for Foo { }
-
-            // In this test, we first try to solve to solve `T:
-            // NonEnumerable` but then we discover it's
-            // non-enumerable, and so we push it off for later. Then
-            // we try to solve the `T: Enumerable` trait.
-
-            trait Debug1 { }
-            impl<T> Debug1 for T where T: Enumerable, T: NonEnumerable { }
-
-            trait Debug2 { }
-            impl<T> Debug2 for T where T: NonEnumerable, T: Enumerable { }
-        }
-
-        goal {
-            exists<A> { A: Debug1 }
-        } first 10 with max 3 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [?0 := Foo]
-                            constraints: []
-                        }
-                        binders: []
-                    },
-                    ambiguous: false
-                }
-            ]"
-        }
-
-
-        goal {
-            exists<A> { A: Debug2 }
-        } first 10 with max 3 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [?0 := Foo]
-                            constraints: []
-                        }
-                        binders: []
-                    },
-                    ambiguous: false
-                }
-            ]"
-        }
-    }
-}
-
 #[test]
 fn auto_traits_flounder() {
     test! {
@@ -1126,457 +604,6 @@ fn auto_traits_flounder() {
             exists<A> { A: Send }
         } first 10 with max 3 {
             r"Floundered"
-        }
-    }
-}
-
-#[test]
-fn negative_reorder() {
-    test! {
-        program {
-            struct Foo { }
-            struct Bar { }
-
-            trait IsFoo { }
-            impl IsFoo for Foo { }
-
-            trait Enumerable { }
-            impl Enumerable for Foo { }
-            impl Enumerable for Bar { }
-
-            // In this test, we first try to solve to solve `not { T:
-            //  IsFoo }` but then we discover it's
-            // non-enumerable, and so we push it off for later. Then
-            // we try to solve the `T: Enumerable` trait.
-
-            trait Debug1 { }
-            forall<T> {
-                T: Debug1 if T: Enumerable, not { T: IsFoo }
-            }
-
-            trait Debug2 { }
-            forall<T> {
-                T: Debug2 if not { T: IsFoo }, T: Enumerable
-            }
-        }
-
-        goal {
-            exists<A> { A: Debug1 }
-        } first 10 with max 3 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [?0 := Bar]
-                            constraints: []
-                        }
-                        binders: []
-                    },
-                    ambiguous: false
-                }
-            ]"
-        }
-
-
-        goal {
-            exists<A> { A: Debug2 }
-        } first 10 with max 3 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [?0 := Bar]
-                            constraints: []
-                        }
-                        binders: []
-                    },
-                    ambiguous: false
-                }
-            ]"
-        }
-    }
-}
-
-/// Test a tricky case for coinductive handling:
-///
-/// While proving C1, we try to prove C2, which recursively requires
-/// proving C1.  If you are naive, you will assume that C2 therefore
-/// holds -- but this is wrong, because C1 later fails when proving
-/// C3.
-#[test]
-fn coinductive_unsound1() {
-    test! {
-        program {
-            trait C1orC2 { }
-
-            #[coinductive]
-            trait C1 { }
-
-            #[coinductive]
-            trait C2 { }
-
-            #[coinductive]
-            trait C3 { }
-
-            forall<T> {
-                T: C1 if T: C2, T: C3
-            }
-
-            forall<T> {
-                T: C2 if T: C1
-            }
-
-            forall<T> {
-                T: C1orC2 if T: C1
-            }
-
-            forall<T> {
-                T: C1orC2 if T: C2
-            }
-        }
-
-        goal {
-            forall<X> { X: C1orC2 }
-        } first 10 with max 3 {
-            r"[]"
-        }
-    }
-}
-
-/// The only difference between this test and `coinductive_unsound1`
-/// is the order of the final `forall` clauses.
-#[test]
-fn coinductive_unsound2() {
-    test! {
-        program {
-            trait C1orC2 { }
-
-            #[coinductive]
-            trait C1 { }
-
-            #[coinductive]
-            trait C2 { }
-
-            #[coinductive]
-            trait C3 { }
-
-            forall<T> {
-                T: C1 if T: C2, T: C3
-            }
-
-            forall<T> {
-                T: C2 if T: C1
-            }
-
-            forall<T> {
-                T: C1orC2 if T: C2
-            }
-
-            forall<T> {
-                T: C1orC2 if T: C1
-            }
-        }
-
-        goal {
-            forall<X> { X: C1orC2 }
-        } first 10 with max 3 {
-            r"[]"
-        }
-    }
-}
-
-// See rust-lang/chalk#280
-#[test]
-fn max_size_infinite_loop_regression() {
-    test! {
-        program {
-            trait FnOnce<Args> {
-                type Output;
-            }
-
-            trait Try {
-                type Ok;
-                type Error;
-            }
-
-            struct Tuple<A, B> { }
-
-            trait ParallelIterator {
-                type Item;
-            }
-        }
-
-        //fn try_reduce_with<PI, R, T>(pi: PI, reduce_op: R) -> Option<T>
-        //    where
-        //        PI: ParallelIterator<Item = T>,
-        //        R: FnOnce(T::Ok) -> T,
-        //        T: Try,
-        //    {
-        //        pi.drive_unindexed()
-        //    }
-        //
-        // where `drive_unindexed` is a method in `ParallelIterator`:
-        //
-        // fn drive_unindexed(self) -> ();
-
-        goal {
-            forall<PI, R, T> {
-                if (
-                    PI: ParallelIterator<Item = T>;
-                    R: FnOnce<Tuple< <T as Try>::Ok, <T as Try>::Ok >>;
-                    T: Try
-                ) {
-                    PI: ParallelIterator
-                }
-            }
-        } first 1 with max 4 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [],
-                            constraints: []
-                        }
-                        binders: []
-                    }
-                    ambiguous: false
-                }
-           ]"
-        }
-
-        goal {
-            forall<PI, R, T> {
-                if (
-                    PI: ParallelIterator<Item = T>;
-                    R: FnOnce<Tuple< <T as Try>::Ok, <T as Try>::Ok >>;
-                    T: Try
-                ) {
-                    PI: ParallelIterator
-                }
-            }
-        } first 1 with max 10 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: [],
-                            constraints: []
-                        }
-                        binders: []
-                    }
-                    ambiguous: false
-                }
-           ]"
-        }
-    }
-}
-
-#[test]
-fn coinductive_multicycle1() {
-    test! {
-        program {
-            trait Any { }
-
-            #[coinductive]
-            trait C1 { }
-
-            #[coinductive]
-            trait C2 { }
-
-            #[coinductive]
-            trait C3 { }
-
-            forall<T> {
-                T: C1 if T: C2
-            }
-
-            forall<T> {
-                T: C2 if T: C3
-            }
-
-            forall<T> {
-                T: C3 if T: C1
-            }
-
-            forall<T> {
-                T: Any if T: C3
-            }
-
-            forall<T> {
-                T: Any if T: C2
-            }
-
-            forall<T> {
-                T: Any if T: C1
-            }
-        }
-
-        goal {
-            forall<X> { X: Any }
-        } first 10 with max 3 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: []
-                            constraints: []
-                        }
-                        binders: []
-                    },
-                    ambiguous: false
-                }
-            ]"
-        }
-    }
-}
-
-#[test]
-fn coinductive_multicycle2() {
-    test! {
-        program {
-            trait Any { }
-
-            #[coinductive]
-            trait C1 { }
-
-            #[coinductive]
-            trait C2 { }
-
-            #[coinductive]
-            trait C3 { }
-
-            forall<T> {
-                T: C1 if T: C2
-            }
-
-            forall<T> {
-                T: C2 if T: C3
-            }
-
-            forall<T> {
-                T: C3 if T: C1, T: C2
-            }
-
-            forall<T> {
-                T: Any if T: C1
-            }
-        }
-
-        goal {
-            forall<X> { X: Any }
-        } first 10 with max 3 {
-            r"[
-                CompleteAnswer {
-                    subst: Canonical {
-                        value: ConstrainedSubst {
-                            subst: []
-                            constraints: []
-                        }
-                        binders: []
-                    },
-                    ambiguous: false
-                }
-            ]"
-        }
-    }
-}
-
-#[test]
-fn coinductive_multicycle3() {
-    test! {
-        program {
-            trait Any { }
-
-            #[coinductive]
-            trait C1 { }
-
-            #[coinductive]
-            trait C2 { }
-
-            #[coinductive]
-            trait C3 { }
-
-            trait C4 { }
-
-            forall<T> {
-                T: C1 if T: C2
-            }
-
-            forall<T> {
-                T: C2 if T: C3, T: C4
-            }
-
-            forall<T> {
-                T: C3 if T: C1
-            }
-
-            forall<T> {
-                T: Any if T: C3
-            }
-
-            forall<T> {
-                T: Any if T: C2
-            }
-
-            forall<T> {
-                T: Any if T: C1
-            }
-        }
-
-        goal {
-            forall<X> { X: Any }
-        } first 10 with max 3 {
-            r"[]"
-        }
-    }
-}
-
-#[test]
-fn coinductive_multicycle4() {
-    test! {
-        program {
-            trait Any { }
-
-            #[coinductive]
-            trait C1 { }
-
-            #[coinductive]
-            trait C2 { }
-
-            #[coinductive]
-            trait C3 { }
-
-            trait C4 { }
-
-            forall<T> {
-                T: C1 if T: C2
-            }
-
-            forall<T> {
-                T: C2 if T: C3
-            }
-
-            forall<T> {
-                T: C3 if T: C1, T: C4
-            }
-
-            forall<T> {
-                T: Any if T: C3
-            }
-
-            forall<T> {
-                T: Any if T: C2
-            }
-
-            forall<T> {
-                T: Any if T: C1
-            }
-        }
-
-        goal {
-            forall<X> { X: Any }
-        } first 10 with max 3 {
-            r"[]"
         }
     }
 }
